Set up `_start` function for the web target
Before zig let use compile a static library which would export _start,
however more recent versions have changed the behaviour to output an
archive (`.a`) file. As a dynamic library the main function is never
called from anywhere, and thus is excluded from the library.

This commit fixes that by having the user add a `comptime` block that
calls `seizer.run`, which will ensure that an entry point is called
from somewhere.