refactor: give seizer control of main loop
Inspired by [Hypermedia Systems][], I'm refactoring seizer to be a way
to write hypermedia applications. For a while I've been wanting to to
give seizer control of the main loop, and have an init function that
sets everything up. This commit takes it a step further and has the
init function declaritively describe itself.

[Hypermedia Systems]: https://hypermedia.systems/