refactor: Remove Context and make comptime-known Platform
Initially I wanted the platform to be something that could be figured out at
runtime. This is especially relevant for the Linux backend that has separate
codepaths for running on a bare EGL system or a system running wayland. However
the runtime-known approach led to passing around a Context pointer that wasn't
progated everywhere, leading to some pretty ugly hacks in games/examples that
were using seizer.

This commit swaps `seizer.Context` for `seizer.Platform`. `seizer.Platform` is
comptime known, and contains the list of functions neccessary to make a usable
platform. This includes the main function, a function to create windows, and
some functions for asynchronous IO. Platforms are expected to use global state
to implement functionality. The end result is a cleaner API applications using
seizer.

This does mean that seizer loses some of its transparency for developers. I
think this trade off is worth it, given a goal of writing games that can run
anywhere, from little Linux handhelds to webpages.

I also deleted some glfw code that has not been used for quite some time.