feat: add PanZoom and Image elements, add ui_view_image example
This is bundled with a change to the way transforms are handled in the
ui. Instead of returning what the current transform is, Elements
implement a function that allows then to get the transform and rectangle
of a child element.

The way I'm viewing this change is as a reduction in the amount of
caching happening. Caching is hard to get right, I think it makes sense
to add a function pointer instead of having a field with redundant,
potentially invalid, state.