feat: add renderdoc support
Normally, renderdoc can just handle launching a program and wrapping the
graphics API. However, since we are avoiding the system libraries and
using dmabufs to submit frames to wayland, renderdoc has no way of
knowing when a frame has begun or ended. Integrating renderdoc let's us
notify renderdoc when we submit a frame.

Compared to most programs that render to the screen, seizer essentially
acts as a headless renderer, that just so happens to show something to
the screen.