fix: vulkan: make transform matrix from window space to clip space
It confused me for a bit, but I think I'm a bit better at understanding
coordinate space transformations now. At it's core, the problem boiled
down to OpenGL and Vulkan having different coordinate conventions[0]. The
orthographic projection matrix designed for OpenGL does not work for
Vulkan. To replace it I scale the vertex coordinates down to the 0..2
coordinate space, and then I translate the vertex by <-1,-1>.

I think I prefer the look of the Vulkan transforms to the monolithic
OpenGL `orthographic` transform, as it makes it clear what the components
are and how they can be modified.

[0]: https://www.kdab.com/projection-matrices-with-vulkan-part-1/