fix vulkan fd export by using the proper memory offset
Part of the problem was that I didn't realize that I was exporting an fd
for the vk.DeviceMemory, not for the vk.Image. The second problem was
that the problem was almost impossible to see; I noticed that there was
some additional choppiness, but I had no idea why.